+++ src/transactions/rule_verifing/RuleVerifier.h
@@ -34,7 +34,7 @@ protected:
     isAssetMatches(AssetFields expectedAsset, AssetFields actualAsset) const;
 
     bool
-    isTasksMatch(uint32_t const expectedTasks, uint32_t const actualTasks) const;
+    isTasksMatch(uint64_t const expectedTasks, uint64_t const actualTasks) const;
 };
 
 } // namespace stellar
